pyproject.toml
[tool.poetry]
name = "linkedin-api-integration"
version = "0.1.0"
description = ""
authors = ["Bruno de Oliveira <obruno.oliveira@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
requests = "^2.32.3"
linkedin = "^0.1.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
